+++ lib/src/radio/theme/default_sbb_radio_theme_data.dart
@@ -1,14 +1,15 @@
-import 'package:flutter/cupertino.dart';
-import 'package:sbb_design_system_mobile/sbb_design_system_mobile.dart';
+import 'package:flutter/material.dart';
 
-/// The default checkbox theme is specified using the
+import '../../../sbb_design_system_mobile.dart';
+
+/// The default radio theme is specified using the
 /// [Figma design specs](https://www.figma.com/design/ZBotr4yqcEKqqVEJTQfSUa/Design-System-Mobile?node-id=123-234&p=f).
-class DefaultSBBCheckboxThemeData extends SBBCheckboxThemeData {
-  DefaultSBBCheckboxThemeData(SBBBaseStyle baseStyle)
+class DefaultSBBRadioThemeData extends SBBRadioThemeData {
+  DefaultSBBRadioThemeData(SBBBaseStyle baseStyle)
     : super(
-        style: SBBCheckboxStyle(
+        style: SBBRadioStyle(
           fillColor: WidgetStatePropertyAll(baseStyle.themeValue(SBBColors.white, SBBColors.charcoal)),
-          checkColor: WidgetStateProperty.fromMap(<WidgetStatesConstraint, Color?>{
+          innerCircleColor: WidgetStateProperty.fromMap(<WidgetStatesConstraint, Color?>{
             WidgetState.disabled: baseStyle.themeValue(SBBColors.granite, SBBColors.graphite),
             WidgetState.any: baseStyle.primaryColor,
           }),