+++ lib/src/theme/sbb_theme.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:sbb_design_system_mobile/sbb_design_system_mobile.dart';
 import 'package:sbb_design_system_mobile/src/button/theme/sbb_button_style_x.dart';
 import 'package:sbb_design_system_mobile/src/checkbox/theme/default_sbb_checkbox_theme_data.dart';
+import 'package:sbb_design_system_mobile/src/radio/theme/default_sbb_radio_theme_data.dart';
 
 import '../button/theme/default_button_themes.dart';
 import '../status/theme/default_sbb_status_theme.dart';
@@ -18,6 +19,7 @@ class SBBTheme {
     SBBSecondaryButtonThemeData? secondaryButtonTheme,
     SBBTertiaryButtonThemeData? tertiaryButtonTheme,
     SBBCheckboxThemeData? checkboxTheme,
+    SBBRadioThemeData? radioTheme,
     SBBControlStyles? controlStyles,
     SBBHeaderBoxStyle? headerBoxStyle,
     SBBGroupStyle? groupStyle,
@@ -32,6 +34,7 @@ class SBBTheme {
     secondaryButtonTheme: secondaryButtonTheme,
     tertiaryButtonTheme: tertiaryButtonTheme,
     checkboxTheme: checkboxTheme,
+    radioTheme: radioTheme,
     controlStyles: controlStyles,
     headerBoxStyle: headerBoxStyle,
     groupStyle: groupStyle,
@@ -47,6 +50,7 @@ class SBBTheme {
     SBBSecondaryButtonThemeData? secondaryButtonTheme,
     SBBTertiaryButtonThemeData? tertiaryButtonTheme,
     SBBCheckboxThemeData? checkboxTheme,
+    SBBRadioThemeData? radioTheme,
     SBBControlStyles? controlStyles,
     SBBHeaderBoxStyle? headerBoxStyle,
     SBBGroupStyle? groupStyle,
@@ -61,6 +65,7 @@ class SBBTheme {
     secondaryButtonTheme: secondaryButtonTheme,
     tertiaryButtonTheme: tertiaryButtonTheme,
     checkboxTheme: checkboxTheme,
+    radioTheme: radioTheme,
     controlStyles: controlStyles,
     headerBoxStyle: headerBoxStyle,
     groupStyle: groupStyle,
@@ -77,6 +82,7 @@ class SBBTheme {
     SBBSecondaryButtonThemeData? secondaryButtonTheme,
     SBBTertiaryButtonThemeData? tertiaryButtonTheme,
     SBBCheckboxThemeData? checkboxTheme,
+    SBBRadioThemeData? radioTheme,
     SBBControlStyles? controlStyles,
     SBBHeaderBoxStyle? headerBoxStyle,
     SBBGroupStyle? groupStyle,
@@ -99,6 +105,9 @@ class SBBTheme {
 
     final defaultCheckboxTheme = DefaultSBBCheckboxThemeData(mergedBaseStyle);
     final mergedCheckboxTheme = defaultCheckboxTheme.merge(checkboxTheme);
+
+    final defaultRadioTheme = DefaultSBBRadioThemeData(mergedBaseStyle);
+    final mergedRadioTheme = defaultRadioTheme.merge(radioTheme);
 
     final defaultControlStyles = SBBControlStyles.$default(baseStyle: mergedBaseStyle);
     final mergedControlStyles = controlStyles.merge(defaultControlStyles);
@@ -125,6 +134,7 @@ class SBBTheme {
       secondaryButtonTheme: mergedSecondaryButtonTheme,
       tertiaryButtonTheme: mergedTertiaryButtonTheme,
       checkboxTheme: mergedCheckboxTheme,
+      radioTheme: mergedRadioTheme,
       controlStyles: mergedControlStyles,
       headerBoxStyle: mergedHeaderBoxStyle,
       groupStyle: mergedGroupStyle,
@@ -141,6 +151,7 @@ class SBBTheme {
     required SBBSecondaryButtonThemeData secondaryButtonTheme,
     required SBBTertiaryButtonThemeData tertiaryButtonTheme,
     required SBBCheckboxThemeData checkboxTheme,
+    required SBBRadioThemeData radioTheme,
     required SBBControlStyles controlStyles,
     required SBBHeaderBoxStyle headerBoxStyle,
     required SBBGroupStyle groupStyle,
@@ -173,6 +184,7 @@ class SBBTheme {
         secondaryButtonTheme,
         tertiaryButtonTheme,
         checkboxTheme,
+        radioTheme,
         controlStyles,
         headerBoxStyle,
         groupStyle,